requirements.txt
streamlit==1.32.0
torch==2.1.0
torchvision==0.16.0
efficientnet_pytorch==0.7.1
opencv-python-headless==4.8.0
albumentations==1.3.1
numpy==1.24.3
pillow==10.0.0
matplotlib==3.7.1
onnx==1.15.0
onnxruntime==1.16.3